merge dev into master for nmsprime 2.5.0 release

+++ resources/views/Generic/index.blade.php
@@ -10,21 +10,21 @@
 --}}
 
 <?php
-	$blade_type = 'index_list';
+    $blade_type = 'index_list';
 ?>
 
 @extends ('Layout.split84-nopanel')
 
 @section('content_top')
-	<li class="active">
-		<a href="{{route($route_name.'.index')}}">
-		    {!! $model->view_icon().' '.$headline !!}
-		</a>
-	</li>
+    <li class="active">
+        <a href="{{route($route_name.'.index')}}">
+            {!! $model->view_icon().' '.$headline !!}
+        </a>
+    </li>
 @stop
 
 @section('content_left')
-	{{-- Headline: means icon followed by headline --}}
+    {{-- Headline: means icon followed by headline --}}
     <div class="col-md-12">
         <div class="row m-b-25">
             <div class="col">
@@ -50,8 +50,8 @@
             <div class="align-self-end m-r-30">
                 @if ($delete_allowed)
                     <button type="submit" class="btn btn-outline-danger m-b-10 float-right" style="simple" data-toggle="tooltip" data-delay='{"show":"250"}' data-placement="top"
-                    title="{{ \App\Http\Controllers\BaseViewController::translate_view('Delete', 'Button' ) }}" form="IndexForm" name="_delete">
-                            <i class="fa fa-trash-o fa-2x" aria-hidden="true"></i>
+                        title="{{ \App\Http\Controllers\BaseViewController::translate_view('Delete', 'Button' ) }}" form="IndexForm" name="_delete">
+                        <i class="fa fa-trash-o fa-2x" aria-hidden="true"></i>
                     </button>
                 @endif
             </div>
@@ -63,11 +63,19 @@
         </div>
     </div>
 
-	@include('Generic.above_infos')
+    @include('Generic.above_infos')
 
     {{-- database entries inside a form with checkboxes to be able to delete one or more entries --}}
-    {{ Form::open(array('route' => array($route_name.'.destroy', 0), 'method' => 'delete', 'id' => 'IndexForm')) }}
+    @if ($delete_allowed)
+        {{ Form::open(array('route' => array($route_name.'.destroy', 0), 'method' => 'delete', 'id' => 'IndexForm')) }}
+    @endif
     {{-- INIT DT --}}
+
+    @php
+        $methodExists = method_exists($model, 'view_index_label');
+        $indexTableInfo = $methodExists ? $model->view_index_label() : [];
+    @endphp
+
     <table class="table table-hover datatable table-bordered d-table" id="datatable">
         {{-- Get Headerdata and translate with translation files --}}
         <thead> {{-- TABLE HEADER --}}
@@ -80,15 +88,15 @@
                     </th>
                 @endif
                 {{-- Get Header if possible with new Format - for Backwards compatibility old one stays --}}
-                @if (isset($model) && method_exists( BaseController::get_model_obj() , 'view_index_label' ) && is_array($model->view_index_label()) && isset($model->view_index_label()['index_header']))
-                    @foreach ($model->view_index_label()['index_header'] as $field)
+                @if (isset($model) && $methodExists && is_array($indexTableInfo) && isset($indexTableInfo['index_header']))
+                    @foreach ($indexTableInfo['index_header'] as $field)
                         <th class="content" style="text-align:center; vertical-align:middle;">{{ trans('dt_header.'.$field).' ' }}
-                        @if ((!empty($model->view_index_label()['disable_sortsearch'])) && ($model->view_index_label()['disable_sortsearch'] == [$field => 'false']))
-                            <i class="fa fa-info-circle text-info" data-trigger="hover" data-container="body" data-toggle="tooltip" data-placement="top" data-delay='{"show":"250"}'
-                            data-original-title="{{trans('helper.SortSearchColumn')}}"></i>
-                        @elseif (!empty($model->view_index_label()['help'][$field]))
-                            <i class="fa fa-info-circle text-info" data-trigger="hover" data-container="body" data-toggle="tooltip" data-placement="top" data-delay='{"show":"250"}'
-                            data-original-title="{{trans('helper.'.$model->view_index_label()['help'][$field])}}"></i>
+                        @if ((!empty($indexTableInfo['disable_sortsearch'])) && ($indexTableInfo['disable_sortsearch'] == [$field => 'false']))
+                            <i class="fa fa-info-circle text-info" data-trigger="hover" data-container="body" data-toggle="tooltip" data-placement="top"
+                                data-delay='{"show":"250"}' data-original-title="{{trans('helper.SortSearchColumn')}}"></i>
+                        @elseif (!empty($indexTableInfo['help'][$field]))
+                            <i class="fa fa-info-circle text-info" data-trigger="hover" data-container="body" data-toggle="tooltip" data-placement="top"
+                                data-delay='{"show":"250"}' data-original-title="{{trans('helper.'.$indexTableInfo['help'][$field])}}"></i>
                         @endif
                         </th>
                     @endforeach
@@ -98,14 +106,14 @@
         <tbody> {{-- Table DATA --}}
         </tbody>
         <tfoot> {{-- TABLE FOOTER--}}
-        @if (isset($model) && method_exists( BaseController::get_model_obj() , 'view_index_label' ))
+        @if (isset($model) && $methodExists)
             <tr>
                 <th></th>  {{-- Responsive Column --}}
                 @if (isset($delete_allowed) && $delete_allowed == true)
                     <th></th> {{-- Checkbox Column if delete is allowed --}}
                 @endif
-                @foreach ($model->view_index_label()['index_header'] as $field)
-                    @if ((!empty($model->view_index_label()['disable_sortsearch'])) && ( array_has( $model->view_index_label()['disable_sortsearch'] , $field) ) )
+                @foreach ($indexTableInfo['index_header'] as $field)
+                    @if ((!empty($indexTableInfo['disable_sortsearch'])) && ( array_has( $indexTableInfo['disable_sortsearch'] , $field) ) )
                         <th></th>
                     @else
                         <th class="searchable"></th>
@@ -115,7 +123,9 @@
         @endif
         </tfoot>
     </table>
-    {{ Form::close() }}
+    @if ($delete_allowed)
+        {{ Form::close() }}
+    @endif
 @stop
 
 @section('javascript')
@@ -141,6 +151,7 @@ $(document).ready(function() {
         autoWidth: false, {{-- Option to ajust Table to Width of container --}}
         dom: 'lBfrtip', {{-- sets order and what to show  --}}
         stateSave: true, {{-- Save Search Filters and visible Columns --}}
+        stateDuration: 0, // 60 * 60 * 24, {{-- Time the State is used - set to 24h --}}
         lengthMenu:  [ [10, 25, 100, 250, 500, -1], [10, 25, 100, 250, 500, "{{ trans('view.jQuery_All') }}" ] ], {{-- Filter to List # Datasets --}}
         {{-- Responsive Column --}}
         columnDefs: [],
@@ -172,7 +183,7 @@ $(document).ready(function() {
             processing: true, {{-- show loader--}}
             serverSide: true, {{-- enable Serverside Handling--}}
             deferRender: true,
-            ajax: '{{ isset($route_name) && $route_name!= "Config.index"  ? route($route_name.'.data') : "" }}',
+            ajax: '{{ isset($ajax_route_name) && $route_name != "Config.index" ? route($ajax_route_name) : "" }}',
             {{-- generic Col Header generation --}}
                 @include('datatables.genericColHeader')
         @endif
